refactored custom cli
