+++ pyproject.toml
@@ -27,6 +27,7 @@ Jinja2 = "^3.0.1"
 questionary = "^1.9.0"
 requests = "^2.25.1"
 pandas = "^1.2.4"
+anndata = "^0.7.6"
 
 [tool.poetry.dev-dependencies]
 pytest = "^6.2.3"
